+++ packages/in_app_purchase/in_app_purchase_platform_interface/test/src/types/product_details_test.dart
@@ -17,6 +17,7 @@ void main() {
           description: 'description',
           price: '13.37',
           currencyCode: 'USD',
+          currencySymbol: r'$',
           rawPrice: 13.37);
 
       expect(productDetails.id, 'id');
@@ -24,6 +25,7 @@ void main() {
       expect(productDetails.description, 'description');
       expect(productDetails.rawPrice, 13.37);
       expect(productDetails.currencyCode, 'USD');
+      expect(productDetails.currencySymbol, r'$');
     });
   });