+++ packages/in_app_purchase/in_app_purchase_android/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.1.4+3
+
+- Updated installation instructions in README.
+
+## 0.1.4+2
+
+* Added price currency symbol to SkuDetailsWrapper.
+
 ## 0.1.4+1
 
 * Fixed typos.