+++ packages/in_app_purchase/in_app_purchase_platform_interface/CHANGELOG.md
@@ -1,6 +1,10 @@
-## 1.0.2
+## 1.1.1
 
 * Added new `PurchaseStatus` named `canceled` to distinguish between an error and user cancellation. 
+
+## 1.1.0
+
+* Added `currencySymbol` in ProductDetails.
 
 ## 1.0.1